+++ TelnetNegotiationCore.TestServer/KestrelMockServer.cs
@@ -1,12 +1,10 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
-using System.Threading;
 using System.Threading.Tasks;
 using TelnetNegotiationCore.Interpreters;
 using TelnetNegotiationCore.Models;
 using Microsoft.AspNetCore.Connections;
-using System.IO.Pipelines;
 using Microsoft.Extensions.Logging;
 using TelnetNegotiationCore.Handlers;
 using TelnetNegotiationCore.Builders;
@@ -22,18 +20,6 @@ namespace TelnetNegotiationCore.TestServer
 		{
 			Console.OutputEncoding = Encoding.UTF8;
 			_logger = logger;
-		}
-
-		private async ValueTask WriteToOutputStreamAsync(ReadOnlyMemory<byte> arg, PipeWriter writer)
-		{
-			try
-			{
-				await writer.WriteAsync(arg, CancellationToken.None);
-			}
-			catch (ObjectDisposedException ode)
-			{
-				_logger.LogError(ode, "Stream has been closed");
-			}
 		}
 
 		public ValueTask SignalGMCPAsync((string module, string writeback) val)
@@ -87,11 +73,10 @@ namespace TelnetNegotiationCore.TestServer
 					Sendable_Variables = () => ["ROOM"],
 				});
 
-				var telnet = await new TelnetInterpreterBuilder()
+				var (telnet, readTask) = await new TelnetInterpreterBuilder()
 				.UseMode(TelnetInterpreter.TelnetMode.Server)
 				.UseLogger(_logger)
 				.OnSubmit(WriteBackAsync)
-				.OnNegotiation(x => WriteToOutputStreamAsync(x, connection.Transport.Output))
 				.AddPlugin<NAWSProtocol>()
 					.OnNAWS(SignalNAWSAsync)
 				.AddPlugin<GMCPProtocol>()
@@ -100,51 +85,26 @@ namespace TelnetNegotiationCore.TestServer
 					.OnMSDPMessage((t, config) => SignalMSDPAsync(msdpHandler, t, config))
 				.AddPlugin<MSSPProtocol>()
 					.OnMSSP(SignalMSSPAsync)
+					.WithMSSPConfig(() => new MSSPConfig
+					{
+						Name = "My Telnet Negotiated Server",
+						UTF_8 = true,
+						Gameplay = ["ABC", "DEF"],
+						Extended = new Dictionary<string, dynamic>
+						{
+							{ "Foo",  "Bar"},
+							{ "Baz", (string[])["Moo", "Meow"] }
+						}
+					})
 				.AddPlugin<TerminalTypeProtocol>()
 				.AddPlugin<CharsetProtocol>()
+					.WithCharsetOrder(Encoding.GetEncoding("utf-8"), Encoding.GetEncoding("iso-8859-1"))
 				.AddPlugin<EORProtocol>()
 				.AddPlugin<SuppressGoAheadProtocol>()
-				.BuildAsync();
+				.BuildAndStartAsync(connection.Transport);
 
-			// Configure MSSP
-			var mssp = telnet.PluginManager!.GetPlugin<MSSPProtocol>();
-			if (mssp != null)
-			{
-				mssp.SetMSSPConfig(() => new MSSPConfig
-				{
-					Name = "My Telnet Negotiated Server",
-					UTF_8 = true,
-					Gameplay = ["ABC", "DEF"],
-					Extended = new Dictionary<string, dynamic>
-					{
-						{ "Foo",  "Bar"},
-						{ "Baz", (string[])["Moo", "Meow"] }
-					}
-				});
-			}
+				await readTask;
 
-			// Configure Charset
-			var charset = telnet.PluginManager!.GetPlugin<CharsetProtocol>();
-			if (charset != null)
-				charset.CharsetOrder = [Encoding.GetEncoding("utf-8"), Encoding.GetEncoding("iso-8859-1")];
-
-			while (true)
-			{
-					var result = await connection.Transport.Input.ReadAsync();
-					var buffer = result.Buffer;
-
-					foreach (var segment in buffer)
-					{
-						await telnet.InterpretByteArrayAsync(segment);
-					}
-
-					if (result.IsCompleted)
-					{
-						break;
-					}
-
-					connection.Transport.Input.AdvanceTo(buffer.End);
-				}
 				_logger.LogInformation("{ConnectionId} disconnected", connection.ConnectionId);
 			}
 		}